Do not clobber numpy array in grib_set_double_array().
This is to work around a bug when eccodes is used with numpy >= 1.17.4.
Overwriting an existing numpy array variable with a FFI ctype object causes
invalid pointer derefs, which results in a variety of bad outcomes, ranging
from segfaults to invalid data.

The workaround is simply to use a separate working variable to hold the numpy
array in grib_set_double_array(), and avoid overwriting it with the ctype
object.  The workaround has been tested against the following numpy versions:
- 1.17.3
- 1.17.4
- 1.17.5
- 1.18.0
- 1.19.0